refactor: 스프린트10
