+++ src/main/java/com/sprint/mission/discodeit/controller/AuthController.java
@@ -2,9 +2,9 @@ package com.sprint.mission.discodeit.controller;
 
 
 import com.sprint.mission.discodeit.common.TokenUtil;
-import com.sprint.mission.discodeit.dto.data.JwtDto;
-import com.sprint.mission.discodeit.dto.data.JwtInformation;
-import com.sprint.mission.discodeit.dto.data.UserDto;
+import com.sprint.mission.discodeit.dto.model.JwtDto;
+import com.sprint.mission.discodeit.dto.model.JwtInformation;
+import com.sprint.mission.discodeit.dto.model.UserDto;
 import com.sprint.mission.discodeit.dto.request.RoleUpdateRequest;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.mapper.UserMapper;
