+++ src/main/java/com/sprint/mission/discodeit/dto/model/UserDto.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.data;
+package com.sprint.mission.discodeit.dto.model;
 
 import java.util.UUID;
 import lombok.Builder;
