+++ src/main/java/com/sprint/mission/discodeit/mapper/BinaryContentMapper.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.mapper;
 
-import com.sprint.mission.discodeit.dto.data.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.model.BinaryContentDto;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import java.util.List;
 import org.mapstruct.Mapper;
