+++ src/main/java/com/sprint/mission/discodeit/mapper/ReadStatusMapper.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.mapper;
 
-import com.sprint.mission.discodeit.dto.data.ReadStatusDto;
+import com.sprint.mission.discodeit.dto.model.ReadStatusDto;
 import com.sprint.mission.discodeit.entity.ReadStatus;
 import java.util.List;
 import org.mapstruct.Mapper;
