+++ src/main/java/com/sprint/mission/discodeit/security/DiscodeitUserDetails.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.security;
 
-import com.sprint.mission.discodeit.dto.data.UserDto;
+import com.sprint.mission.discodeit.dto.model.UserDto;
 import com.sprint.mission.discodeit.common.Role;
 import java.util.Collection;
 import java.util.List;
