+++ src/main/java/com/sprint/mission/discodeit/security/jwt/JwtAuthenticationFilter.java
@@ -1,7 +1,6 @@
 package com.sprint.mission.discodeit.security.jwt;
 
 import com.nimbusds.jwt.JWTClaimsSet;
-import com.sprint.mission.discodeit.dto.data.UserDto;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -17,7 +16,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
