+++ src/main/java/com/sprint/mission/discodeit/storage/BinaryContentStorage.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.storage;
 
-import com.sprint.mission.discodeit.dto.data.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.model.BinaryContentDto;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.UUID;
